+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bug fixes
 * fix `ignore` macro
 * fix `load` function in node
+* fix `Pair::append` and `append` functions with non pair
 
 ## 0.15.1
 ### Bug fixes